[CLEANUP] Remove unused `disable_headers` flag in `show_request`
Tried to refactor this partial so that it can be reused in the guides.  Unfortunately, it's not as easy to transpose it because the guide html needs are different from the docs ones. (Guide HTML doesn't need all the fields for JSON request, just the ids and other basic ones).